+++ frontend/src/features/map_planning/components/timeline/Timeline.stories.tsx
@@ -1,8 +1,8 @@
-import { Timeline } from './Timeline';
+import TimelineDatePicker from './TimelineDatePicker';
 import type { Meta, StoryObj } from '@storybook/react';
 
-const meta: Meta<typeof Timeline> = {
-  component: Timeline,
+const meta: Meta<typeof TimelineDatePicker> = {
+  component: TimelineDatePicker,
   argTypes: {
     onSelectDate: { action: 'onSelectDate' },
   },
@@ -10,7 +10,7 @@ const meta: Meta<typeof Timeline> = {
 
 export default meta;
 
-type Story = StoryObj<typeof Timeline>;
+type Story = StoryObj<typeof TimelineDatePicker>;
 
 export const Default: Story = {
   args: {},